Added float8 support. Cleaned up C++ source code file names and simplified type checking
