+++ src/csrc/gnn/gat/mlp_flat_bwd.h
@@ -8,7 +8,7 @@
 /* Authors: Ramanarayan Mohanty, Sasikanth Avancha (Intel Corp.)
  ******************************************************************************/
 
-RECORD_FUNCTION("gemm_attn_bwd", std::vector<c10::IValue>());
+RECORD_FUNCTION("fused_mlp_flat_bwd", std::vector<c10::IValue>());
 
 at::Tensor t_in, t_wt;
 
@@ -16,12 +16,12 @@ int i = 0;
 
 const int threads = omp_get_max_threads();
 
-auto t_grad_out = inputs[i++].contiguous();
+auto t_grad_out = inputs[i++].contiguous(); //[N, HF]
 
-t_in = inputs[i++];
-t_wt = inputs[i++];
+t_in = inputs[i++]; // [N, HF]
+t_wt = inputs[i++]; // [nk, nc, bc, bk]
 
-auto in_sizes = t_in.sizes();
+auto in_sizes = t_in.sizes(); // [N C]
 auto N = in_sizes[0];
 
 auto bn = align;
@@ -34,7 +34,7 @@ auto C = in_sizes[1];
 auto nk = wt_sizes[0];
 auto nc = wt_sizes[1];
 auto bc = wt_sizes[2];
-if (t_wt.dtype() == at::kBFloat16)
+if (t_wt.dim() == 5)
   bc = bc * wt_sizes[4];
 auto bk = wt_sizes[3];
 
@@ -44,15 +44,15 @@ auto bcp = bc;
 auto remp = rem;
 
 auto K = nk * bk;
-
-if (t_in.dtype() == at::kBFloat16) {
-  bnp = bn + bn % 2;
-  remp = rem + rem % 2;
+auto lp = get_vnni_block_size(t_in.dtype());
+if (t_in.dim() == 5) {
+  bnp = bn + bn % lp;
+  remp = rem + rem % lp;
 }
 
-if (t_wt.dtype() == at::kBFloat16) {
-  bcp = bc + bc % 2;
-  bkp = bk + bk % 2;
+if (t_wt.dim() == 5) {
+  bcp = bc + bc % lp;
+  bkp = bk + bk % lp;
 }
 
 //----------------------------
@@ -62,14 +62,18 @@ const auto input_trans_flag =
 
 auto t_wt_TV = wt_tensor_for_bwd(nk, bk, nc, bc, t_wt);
 
-auto t_grad_in = t_in.new_empty({N, C});
+auto t_grad_in = inp_needs_grad ? t_in.new_empty({N, C}) : t_in.new_empty(0);
 
 auto t_grad_wt = at::empty_like(t_wt);
 at::Tensor t_grad_wt_tmp;
-if (t_wt.dtype() == at::kBFloat16)
+if (t_wt.dim() == 5)
   t_grad_wt_tmp = at::empty({nk, nc, bc, bk});
 else
   t_grad_wt_tmp = t_grad_wt;
+
+at::Tensor t_grad_bias = t_grad_out.new_empty(0);
+if (add_bias)
+  t_grad_bias = t_grad_out.new_empty({nk * bk});
 
 auto grad_out = GetVLAPtr<T>(t_grad_out, {bn, nk, bk});
 auto grad_in = GetVLAPtr<T>(t_grad_in, {bn, nc, bc});
@@ -79,11 +83,13 @@ auto grad_wt = GetVLAPtr<T>(t_grad_wt, {nc, bc* bk});
 auto grad_wt_tmp = GetVLAPtr<float>(t_grad_wt_tmp, {nc, bc* bk});
 
 auto wt_TV = GetVLAPtr<T>(t_wt_TV, {nc, bkp* bc});
+auto grad_bias = GetVLAPtr<T>(t_grad_bias, {bk});
 
 auto in = GetVLAPtr<T>(t_in, {bn, nc, bc}); // flat layout for fp32
 
 auto set_zero_tpp = SCOPEIT(SetZeroTPP<float>(nk * bk), EW_ZERO);
 auto set_zero_col_tpp = SCOPEIT(SetZeroTPP<T>(bn, 1, bkp), EW_ZERO);
+auto grad_bias_tpp = SCOPEIT(GradBiasTPP<T>(bn, bk, K), BIAS);
 auto n2v_tpp = SCOPEIT(
     XformExtTPP<T>(bn, bk, bnp, bk, nk* bk, bk, XformTPP::XFORM_N2V_TPP, true),
     VNNI);
@@ -91,6 +97,8 @@ auto n2v_wt_tpp = SCOPEIT(
     XformExtTPP<T>(bc, bk, bcp, bk, XformTPP::XFORM_N2V_TPP, true),
     VNNI);
 auto cpy_tpp = SCOPEIT(CpyTPP<T>(bn, bk, bk, bkp), EW_COPY);
+auto cvt_tpp = SCOPEIT((ConvertTPP<float, T>(bn, bk, K, K)), EW_COPY);
+auto cvt_f32_tpp = SCOPEIT((ConvertTPP<T, float>(bn, bk, K, K)), EW_COPY);
 auto add_gwt_tpp = SCOPEIT((AddTPP<float, float>(bc, bk)), EW_ADD);
 
 auto brgemm_di_tpp = SCOPEIT(
@@ -124,106 +132,158 @@ auto brgemm_dw_f32_tpp_b1 = SCOPEIT((BrgemmTPP<T, float>(
     16)));
 
 // BF16 del-wt brgemms
-auto brgemm_dw_bf16_tpp =
+auto brgemm_dw_lp_tpp =
     SCOPEIT((BrgemmTPP<
              T,
              float>(bc, bk, bnp, bc* bnp, bk* bnp, bnp, bk, bk, 0.0, 0, 16)));
-auto brgemm_dw_bf16_tpp_b1 =
+auto brgemm_dw_lp_tpp_b1 =
     SCOPEIT((BrgemmTPP<
              T,
              float>(bc, bk, bnp, bc* bnp, bk* bnp, bnp, bk, bk, 1.0, 0, 16)));
 
-//=================================================================================================================================
-//==========================================Grad_bias & Grad_in fused
-//=============================================================
-//=================================================================================================================================
-
 {
-  RECORD_SCOPE(gadi_gemm_attn, {t_grad_out, t_grad_in});
+  RECORD_SCOPE(gm_dbias, {t_grad_out, t_grad_bias});
   {
-    {
-      RECORD_FUNCTION("parallel_for", std::vector<c10::IValue>());
+    RECORD_FUNCTION("parallel_for", std::vector<c10::IValue>());
+    if (add_bias) {
+      tensor_set_zero(nk, bk, t_grad_bias);
+      float* bias_ptrs[threads];
 #pragma omp parallel
       {
-        T tmp[bn][nk][bkp];
+        int tid = omp_get_thread_num();
+        float prv_grad_bias[nk][bk];
+        bias_ptrs[tid] = prv_grad_bias[0];
+        set_zero_tpp(prv_grad_bias[0]);
 
 #pragma omp for
         for (int n = 0; n < nn; n++) {
-          // Grad_in Brgemm computation if bk != bkp
-          if (bk != bkp) {
-            for (int k = 0; k < nk; k++)
-              set_zero_col_tpp(&tmp[0][k][bk]);
-
-            brgemm_di_tpp.config();
-            for (int c = 0; c < nc; c++) {
-              for (int k = 0; k < nk; k++)
-                cpy_tpp(grad_out[n][0][k], tmp[0][k]);
-              brgemm_di_tpp(tmp[0][0], wt_TV[0][c], grad_in[n][0][c], nk, true);
-            }
-            brgemm_di_tpp.release();
-          } else { // Grad_in Brgemm computation if bk == bkp
-            brgemm_di_tpp.config();
-            for (int c = 0; c < nc; c++) {
-              brgemm_di_tpp(
-                  grad_out[n][0][0], wt_TV[0][c], grad_in[n][0][c], nk, true);
-            }
-            brgemm_di_tpp.release();
+          for (int k = 0; k < nk; k++) {
+            grad_bias_tpp(grad_out[n][0][k], prv_grad_bias[k]);
           }
         }
+        omp_reduce_buf(threads, nk * bk, bias_ptrs, grad_bias[0]);
       }
 
       if (rem > 0) {
+        // Grad_bias---------------------------------------------------
         auto grad_out = GetVLAPtr<T>(t_grad_out, {nk, bk});
 
-        // Grad_in-----------------------------------------------------
-        {
-          auto grad_in = GetVLAPtr<T>(t_grad_in, {nc, bc});
-          auto wt_TV = GetVLAPtr<T>(t_wt_TV, {nc, bc * bkp});
+        auto grad_bias_tpp = SCOPEIT(GradBiasTPP<T>(1, bk, K), BIAS);
 
-          auto set_zero_col_tpp = SCOPEIT(SetZeroTPP<T>(rem, 1, bkp), EW_ZERO);
-          auto cpy_tpp = SCOPEIT(CpyTPP<T>(rem, bk, bk, bkp), EW_COPY);
-          auto brgemm_di_tpp = SCOPEIT((BrgemmTPP<T, T>(
-              rem,
-              bc,
-              bkp,
-              bkp,
-              nc * bc * bkp,
-              nk * bkp,
-              bc,
-              nc * bc,
-              0.0,
-              0,
-              nk)));
+        float prv_grad_bias[nk][bk];
+        bias_ptrs[0] = prv_grad_bias[0];
+        set_zero_tpp(prv_grad_bias[0]);
 
-          if (bk != bkp) {
-            T tmp[rem][nk][bkp];
-
-            for (int k = 0; k < nk; k++)
-              set_zero_col_tpp(&tmp[0][k][bk]);
-
-            brgemm_di_tpp.config();
-
-            for (int c = 0; c < nc; c++) {
-              for (int k = 0; k < nk; k++)
-                cpy_tpp(grad_out[nn * bn][k], tmp[0][k]);
-              brgemm_di_tpp(
-                  tmp[0][0], wt_TV[0][c], grad_in[nn * bn][c], nk, true);
-            }
-            brgemm_di_tpp.release();
-          } else { // Grad_in Brgemm computation if bk == bkp
-            brgemm_di_tpp.config();
-            for (int c = 0; c < nc; c++) {
-              brgemm_di_tpp(
-                  grad_out[nn * bn][0],
-                  wt_TV[0][c],
-                  grad_in[nn * bn][c],
-                  nk,
-                  true);
-            }
-            brgemm_di_tpp.release();
+        for (int k = 0; k < nk; k++) {
+          for (int r = 0; r < rem; r++) {
+            grad_bias_tpp(grad_out[nn * bn + r][k], prv_grad_bias[k]);
           }
         }
-      } // rem > 0
+        omp_reduce_buf(1, nk * bk, bias_ptrs, grad_bias[0], true);
+      }
+    }
+  }
+}
+
+if (inp_needs_grad) {
+  RECORD_SCOPE(gmdi_gemm, {t_grad_out, t_grad_in});
+  {
+    if (bk != bkp) {
+      RECORD_FUNCTION("parallel_for", std::vector<c10::IValue>());
+#pragma omp parallel
+      {
+        brgemm_di_tpp.config();
+        T tmp[bn][nk][bkp];
+        for (int k = 0; k < nk; k++)
+          set_zero_col_tpp(&tmp[0][k][bk]);
+
+        int tid = omp_get_thread_num();
+        int threads = omp_get_num_threads();
+        int work = nn * nc;
+        int chunk =
+            (work % threads == 0) ? (work / threads) : (work / threads) + 1;
+        int chunk_start = (tid * chunk < work) ? (tid * chunk) : work;
+        int chunk_end = ((tid + 1) * chunk < work) ? ((tid + 1) * chunk) : work;
+
+        for (int n3c = chunk_start; n3c < chunk_end; n3c++) {
+          int n = n3c / nc;
+          int c = n3c % nc;
+
+          for (int k = 0; k < nk; k++)
+            cpy_tpp(grad_out[n][0][k], tmp[0][k]);
+
+          brgemm_di_tpp(tmp[0][0], wt_TV[0][c], grad_in[n][0][c], nk, true);
+        }
+        brgemm_di_tpp.release();
+      }
+    } else {
+      RECORD_FUNCTION("parallel_for", std::vector<c10::IValue>());
+#pragma omp parallel
+      {
+        int tid = omp_get_thread_num();
+        int threads = omp_get_num_threads();
+        int work = nn * nc;
+        int chunk =
+            (work % threads == 0) ? (work / threads) : (work / threads) + 1;
+        int chunk_start = (tid * chunk < work) ? (tid * chunk) : work;
+        int chunk_end = ((tid + 1) * chunk < work) ? ((tid + 1) * chunk) : work;
+
+        brgemm_di_tpp.config();
+
+        for (int n3c = chunk_start; n3c < chunk_end; n3c++) {
+          int n = n3c / nc;
+          int c = n3c % nc;
+
+          brgemm_di_tpp.config();
+          brgemm_di_tpp(
+              grad_out[n][0][0], wt_TV[0][c], grad_in[n][0][c], nk, true);
+        }
+        brgemm_di_tpp.release();
+      }
+    }
+
+    if (rem > 0) {
+      // Grad_in-----------------------------------------------------
+      auto grad_out = GetVLAPtr<T>(t_grad_out, {nk, bk});
+      auto grad_in = GetVLAPtr<T>(t_grad_in, {nc, bc});
+      auto wt_TV = GetVLAPtr<T>(t_wt_TV, {nc, bc * bkp});
+
+      auto set_zero_col_tpp = SCOPEIT(SetZeroTPP<T>(rem, 1, bkp), EW_ZERO);
+      auto cpy_tpp = SCOPEIT(CpyTPP<T>(rem, bk, bk, bkp), EW_COPY);
+      auto cvt_tpp = SCOPEIT((ConvertTPP<float, T>(rem, bc, C, C)), EW_COPY);
+      auto brgemm_di_tpp = SCOPEIT((BrgemmTPP<T, T>(
+          rem,
+          bc,
+          bkp,
+          bkp,
+          nc * bc * bkp,
+          nk * bkp,
+          bc,
+          nc * bc,
+          0.0,
+          0,
+          nk)));
+
+      brgemm_di_tpp.config();
+
+      if (bk != bkp) {
+        T tmp[rem][nk][bkp];
+
+        for (int k = 0; k < nk; k++) {
+          set_zero_col_tpp(&tmp[0][k][bk]);
+          cpy_tpp(grad_out[nn * bn][k], tmp[0][k]);
+        }
+        for (int c = 0; c < nc; c++) {
+          brgemm_di_tpp(tmp[0][0], wt_TV[0][c], grad_in[nn * bn][c], nk, true);
+        }
+        brgemm_di_tpp.release();
+      } else { // Grad_in Brgemm computation if bk == bkp
+        for (int c = 0; c < nc; c++) {
+          brgemm_di_tpp(
+              grad_out[nn * bn][0], wt_TV[0][c], grad_in[nn * bn][c], nk, true);
+        }
+      }
+      brgemm_di_tpp.release();
     }
   }
 }
@@ -233,15 +293,15 @@ auto trans_tpp = SCOPEIT(
         T>(bn, bc, bc, bnp, nc* bc, bnp, XformTPP::XFORM_XPOSE_TPP, true),
     XPOSE);
 {
-  RECORD_SCOPE(gadw_gemm_attn, {t_in, t_grad_out});
+  RECORD_SCOPE(gmdw_gemm, {t_in, t_grad_out});
   {
-    {
+    if (nn > 0) {
       int upd_n_weight_copies;
       int BF;
 
 #if 1
-      upd_n_weight_copies = nk * nc < 4 * threads ? threads : 1;
-      BF = 32;
+      upd_n_weight_copies = nk * nc < 4 * threads ? threads : threads / 2;
+      BF = 256;
 #else
       BF = atoi(getenv("BF"));
       upd_n_weight_copies = atoi(getenv("UPD_WEIGHT_COPIES"));
@@ -280,8 +340,8 @@ auto trans_tpp = SCOPEIT(
       at::Tensor t_global_tmp_go = at::empty(0);
       at::Tensor t_global_tmp_inT = at::empty(0);
 
-      if (t_grad_out.dtype() == at::kBFloat16 &&
-          t_wt.dtype() == at::kBFloat16) {
+      if (t_grad_out.dim() == 5 &&
+          t_wt.dim() == 5) {
         t_global_tmp_go =
             at::empty({threads, (nn / BF + 1), bnp * bk}, at::kBFloat16);
         t_global_tmp_inT =
@@ -322,14 +382,14 @@ auto trans_tpp = SCOPEIT(
             ? (my_mb_blocks / BF)
             : ((my_mb_blocks / BF) + 1);
 
-        if (t_grad_out.dtype() == at::kBFloat16)
-          brgemm_dw_bf16_tpp_b1.config();
+        if (t_grad_out.dim() == 5)
+          brgemm_dw_lp_tpp_b1.config();
 
         for (int bfn = my_mb_start; bfn < my_mb_end; bfn += mb_block_step) {
           blocks = (bfn + mb_block_step <= my_mb_end) ? mb_block_step
                                                       : my_mb_end - bfn;
           for (int ofm1 = 0; ofm1 < nk; ++ofm1) {
-            if (t_in.dtype() == at::kBFloat16) {
+            if (t_in.dim() == 5) {
               n2v_tpp(
                   blocks,
                   K * bnp,
@@ -342,13 +402,13 @@ auto trans_tpp = SCOPEIT(
                 /* initiaize current work task to zero */
                 setzero_delwt_tpp(grad_wt_priv[team_id][ofm1][ifm1]);
               }
-              if (t_in.dtype() == at::kFloat) {
+              if (t_in.dim() == 4) {
                 brgemm_dw_f32_tpp_b1(
                     in[bfn][0][ifm1],
                     grad_out[bfn][0][ofm1],
                     grad_wt_priv[team_id][ofm1][ifm1],
                     blocks);
-              } else if (t_in.dtype() == at::kBFloat16) {
+              } else if (t_in.dim() == 5) {
                 if (ofm1 == 0)
                   trans_tpp(
                       blocks,
@@ -357,7 +417,7 @@ auto trans_tpp = SCOPEIT(
                       in[bfn][0][ifm1],
                       global_tmp_inT[tid][ifm1 - my_ifm_start][0]);
 
-                brgemm_dw_bf16_tpp_b1(
+                brgemm_dw_lp_tpp_b1(
                     global_tmp_inT[tid][ifm1 - my_ifm_start][0],
                     global_tmp_go[tid][0],
                     grad_wt_priv[team_id][ofm1][ifm1],
@@ -368,8 +428,8 @@ auto trans_tpp = SCOPEIT(
           }
         }
 
-        if (t_grad_out.dtype() == at::kBFloat16)
-          brgemm_dw_bf16_tpp_b1.release();
+        if (t_grad_out.dim() == 5)
+          brgemm_dw_lp_tpp_b1.release();
 
         const int reduce_thr_begin = (tid * reduce_chunksize < reduce_work)
             ? (tid * reduce_chunksize)
@@ -406,7 +466,7 @@ auto trans_tpp = SCOPEIT(
           1.0,
           input_trans_flag,
           1)));
-      auto brgemm_dw_bf16_tpp_b1 = SCOPEITGEMM((BrgemmTPP<T, float>(
+      auto brgemm_dw_lp_tpp_b1 = SCOPEITGEMM((BrgemmTPP<T, float>(
           bc, bk, remp, bc * remp, bk * remp, remp, bk, bk, 1.0, 0, 1)));
       auto n2v_tpp = SCOPEIT(
           XformExtTPP<T>(
@@ -424,7 +484,7 @@ auto trans_tpp = SCOPEIT(
               true),
           XPOSE);
 
-      if (t_wt.dtype() == at::kFloat) {
+      if (t_wt.dim() == 4) {
 #pragma omp parallel for collapse(2)
         for (int k = 0; k < nk; k++) {
           for (int c = 0; c < nc; c++) {
@@ -432,7 +492,7 @@ auto trans_tpp = SCOPEIT(
                 in[nn * bn][c], grad_out[nn * bn][k], grad_wt_tmp[k][c], 1);
           }
         }
-      } else if (t_wt.dtype() == at::kBFloat16) {
+      } else if (t_wt.dim() == 5) {
         T tmp_go[remp * bk], tmp_inT[remp * bc];
 #pragma omp parallel
         {
@@ -445,7 +505,7 @@ auto trans_tpp = SCOPEIT(
           int chunk_end =
               ((tid + 1) * chunk < work) ? ((tid + 1) * chunk) : work;
 
-          brgemm_dw_bf16_tpp_b1.config();
+          brgemm_dw_lp_tpp_b1.config();
 
           for (int kk = chunk_start; kk < chunk_end; kk++) {
             int k = kk / nc;
@@ -453,13 +513,13 @@ auto trans_tpp = SCOPEIT(
 
             n2v_tpp(grad_out[nn * bn][k], tmp_go);
             trans_tpp(in[nn * bn][c], tmp_inT);
-            brgemm_dw_bf16_tpp_b1(tmp_inT, tmp_go, grad_wt_tmp[k][c], 1, true);
+            brgemm_dw_lp_tpp_b1(tmp_inT, tmp_go, grad_wt_tmp[k][c], 1, true);
           }
-          brgemm_dw_bf16_tpp_b1.release();
+          brgemm_dw_lp_tpp_b1.release();
         }
       }
     }
-    if (t_wt.dtype() == at::kBFloat16) {
+    if (t_wt.dim() == 5) {
 #pragma omp parallel for collapse(2)
       for (int k = 0; k < nk; k++) {
         for (int c = 0; c < nc; c++) {
@@ -467,7 +527,18 @@ auto trans_tpp = SCOPEIT(
         }
       }
     }
+    if (nn == 0 and rem == 0) {
+      auto set_zero_tpp = SetZeroTPP<T>(bk, bc);
+      for (int k = 0; k < nk; k++) {
+        for (int c = 0; c < nc; c++) {
+          set_zero_tpp(grad_wt[k][c]);
+        }
+      }
+    }
   }
 }
 
-return {t_grad_in, t_grad_wt};
+if (add_bias)
+  return {t_grad_in, t_grad_wt, t_grad_bias};
+else
+  return {t_grad_in, t_grad_wt};
